feat: update createWarranty service
